# handlers/handlers.go
// Package handlers provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package handlers

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for StateResponseStateInstance.
const (
	Authorized      StateResponseStateInstance = "authorized"
	Blocked         StateResponseStateInstance = "blocked"
	NotAuthorized   StateResponseStateInstance = "notAuthorized"
	PendingPassword StateResponseStateInstance = "pendingPassword"
	Starting        StateResponseStateInstance = "starting"
	Suspended       StateResponseStateInstance = "suspended"
)

// Valid indicates whether the value is a known member of the StateResponseStateInstance enum.
func (e StateResponseStateInstance) Valid() bool {
	switch e {
	case Authorized:
		return true
	case Blocked:
		return true
	case NotAuthorized:
		return true
	case PendingPassword:
		return true
	case Starting:
		return true
	case Suspended:
		return true
	default:
		return false
	}
}

// Defines values for YesNo.
const (
	No  YesNo = "no"
	Yes YesNo = "yes"
)

// Valid indicates whether the value is a known member of the YesNo enum.
func (e YesNo) Valid() bool {
	switch e {
	case No:
		return true
	case Yes:
		return true
	default:
		return false
	}
}

// AccountSettingsResponse defines model for AccountSettingsResponse.
type AccountSettingsResponse struct {
	Avatar *string `json:"avatar,omitempty"`
	ChatId *string `json:"chatId,omitempty"`
	Phone  *string `json:"phone,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Code    string `json:"code"`
	Details string `json:"details"`
}

// SendFileResponse defines model for SendFileResponse.
type SendFileResponse struct {
	IdMessage *string `json:"idMessage,omitempty"`
}

// SendMessageResponse defines model for SendMessageResponse.
type SendMessageResponse struct {
	IdMessage *string `json:"idMessage,omitempty"`
}

// SettingsResponse defines model for SettingsResponse.
type SettingsResponse struct {
	DelaySendMessagesMilliseconds     *int64  `json:"delaySendMessagesMilliseconds,omitempty"`
	IncomingWebhook                   *YesNo  `json:"incomingWebhook,omitempty"`
	MarkIncomingMessagesReaded        *YesNo  `json:"markIncomingMessagesReaded,omitempty"`
	MarkIncomingMessagesReadedOnReply *YesNo  `json:"markIncomingMessagesReadedOnReply,omitempty"`
	OutgoingAPIMessageWebhook         *YesNo  `json:"outgoingAPIMessageWebhook,omitempty"`
	OutgoingMessageWebhook            *YesNo  `json:"outgoingMessageWebhook,omitempty"`
	OutgoingWebhook                   *YesNo  `json:"outgoingWebhook,omitempty"`
	StateWebhook                      *YesNo  `json:"stateWebhook,omitempty"`
	TypeInstance                      *string `json:"typeInstance,omitempty"`
	WebhookUrl                        *string `json:"webhookUrl,omitempty"`
	WebhookUrlToken                   *string `json:"webhookUrlToken,omitempty"`
	Wid                               *string `json:"wid,omitempty"`
}

// StateResponse defines model for StateResponse.
type StateResponse struct {
	StateInstance *StateResponseStateInstance `json:"stateInstance,omitempty"`
}

// StateResponseStateInstance defines model for StateResponse.StateInstance.
type StateResponseStateInstance string

// YesNo defines model for YesNo.
type YesNo string

// ApiToken defines model for ApiToken.
type ApiToken = string

// InstanceId defines model for InstanceId.
type InstanceId = string

// GetAccountSettingsParams defines parameters for GetAccountSettings.
type GetAccountSettingsParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// GetInstanceSettingsParams defines parameters for GetInstanceSettings.
type GetInstanceSettingsParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// GetInstanceStateParams defines parameters for GetInstanceState.
type GetInstanceStateParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// SendFileJSONBody defines parameters for SendFile.
type SendFileJSONBody struct {
	Caption  *string `json:"caption,omitempty"`
	ChatId   string  `json:"chatId"`
	FileName string  `json:"fileName"`
	UrlFile  string  `json:"urlFile"`
}

// SendFileParams defines parameters for SendFile.
type SendFileParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// SendMessageJSONBody defines parameters for SendMessage.
type SendMessageJSONBody struct {
	ChatId  string `json:"chatId"`
	Message string `json:"message"`
}

// SendMessageParams defines parameters for SendMessage.
type SendMessageParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// SendFileJSONRequestBody defines body for SendFile for application/json ContentType.
type SendFileJSONRequestBody SendFileJSONBody

// SendMessageJSONRequestBody defines body for SendMessage for application/json ContentType.
type SendMessageJSONRequestBody SendMessageJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get account settings
	// (GET /api/v1/getAccountSettings)
	GetAccountSettings(w http.ResponseWriter, r *http.Request, params GetAccountSettingsParams)
	// Get instance settings
	// (GET /api/v1/getSettings)
	GetInstanceSettings(w http.ResponseWriter, r *http.Request, params GetInstanceSettingsParams)
	// Get instance state
	// (GET /api/v1/getStateInstance)
	GetInstanceState(w http.ResponseWriter, r *http.Request, params GetInstanceStateParams)
	// Send file to the chat
	// (POST /api/v1/sendFileByURL)
	SendFile(w http.ResponseWriter, r *http.Request, params SendFileParams)
	// Send message to the chat
	// (POST /api/v1/sendMessage)
	SendMessage(w http.ResponseWriter, r *http.Request, params SendMessageParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get account settings
// (GET /api/v1/getAccountSettings)
func (_ Unimplemented) GetAccountSettings(w http.ResponseWriter, r *http.Request, params GetAccountSettingsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get instance settings
// (GET /api/v1/getSettings)
func (_ Unimplemented) GetInstanceSettings(w http.ResponseWriter, r *http.Request, params GetInstanceSettingsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get instance state
// (GET /api/v1/getStateInstance)
func (_ Unimplemented) GetInstanceState(w http.ResponseWriter, r *http.Request, params GetInstanceStateParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Send file to the chat
// (POST /api/v1/sendFileByURL)
func (_ Unimplemented) SendFile(w http.ResponseWriter, r *http.Request, params SendFileParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Send message to the chat
// (POST /api/v1/sendMessage)
func (_ Unimplemented) SendMessage(w http.ResponseWriter, r *http.Request, params SendMessageParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetAccountSettings operation middleware
func (siw *ServerInterfaceWrapper) GetAccountSettings(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetAccountSettingsParams

	// ------------- Required query parameter "instance_id" -------------

	if paramValue := r.URL.Query().Get("instance_id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "instance_id"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "instance_id", r.URL.Query(), &params.InstanceId, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "instance_id", Err: err})
		return
	}

	// ------------- Required query parameter "api_token" -------------

	if paramValue := r.URL.Query().Get("api_token"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "api_token"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "api_token", r.URL.Query(), &params.ApiToken, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "api_token", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAccountSettings(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetInstanceSettings operation middleware
func (siw *ServerInterfaceWrapper) GetInstanceSettings(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetInstanceSettingsParams

	// ------------- Required query parameter "instance_id" -------------

	if paramValue := r.URL.Query().Get("instance_id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "instance_id"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "instance_id", r.URL.Query(), &params.InstanceId, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "instance_id", Err: err})
		return
	}

	// ------------- Required query parameter "api_token" -------------

	if paramValue := r.URL.Query().Get("api_token"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "api_token"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "api_token", r.URL.Query(), &params.ApiToken, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "api_token", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetInstanceSettings(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetInstanceState operation middleware
func (siw *ServerInterfaceWrapper) GetInstanceState(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetInstanceStateParams

	// ------------- Required query parameter "instance_id" -------------

	if paramValue := r.URL.Query().Get("instance_id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "instance_id"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "instance_id", r.URL.Query(), &params.InstanceId, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "instance_id", Err: err})
		return
	}

	// ------------- Required query parameter "api_token" -------------

	if paramValue := r.URL.Query().Get("api_token"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "api_token"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "api_token", r.URL.Query(), &params.ApiToken, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "api_token", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetInstanceState(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SendFile operation middleware
func (siw *ServerInterfaceWrapper) SendFile(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params SendFileParams

	// ------------- Required query parameter "instance_id" -------------

	if paramValue := r.URL.Query().Get("instance_id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "instance_id"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "instance_id", r.URL.Query(), &params.InstanceId, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "instance_id", Err: err})
		return
	}

	// ------------- Required query parameter "api_token" -------------

	if paramValue := r.URL.Query().Get("api_token"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "api_token"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "api_token", r.URL.Query(), &params.ApiToken, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "api_token", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SendFile(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SendMessage operation middleware
func (siw *ServerInterfaceWrapper) SendMessage(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params SendMessageParams

	// ------------- Required query parameter "instance_id" -------------

	if paramValue := r.URL.Query().Get("instance_id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "instance_id"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "instance_id", r.URL.Query(), &params.InstanceId, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "instance_id", Err: err})
		return
	}

	// ------------- Required query parameter "api_token" -------------

	if paramValue := r.URL.Query().Get("api_token"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "api_token"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "api_token", r.URL.Query(), &params.ApiToken, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "api_token", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SendMessage(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/getAccountSettings", wrapper.GetAccountSettings)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/getSettings", wrapper.GetInstanceSettings)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/getStateInstance", wrapper.GetInstanceState)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/sendFileByURL", wrapper.SendFile)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/sendMessage", wrapper.SendMessage)
	})

	return r
}