# models/models.go
// Package models provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package models

// Defines values for StateResponseStateInstance.
const (
	Authorized      StateResponseStateInstance = "authorized"
	Blocked         StateResponseStateInstance = "blocked"
	NotAuthorized   StateResponseStateInstance = "notAuthorized"
	PendingPassword StateResponseStateInstance = "pendingPassword"
	Starting        StateResponseStateInstance = "starting"
	Suspended       StateResponseStateInstance = "suspended"
)

// Valid indicates whether the value is a known member of the StateResponseStateInstance enum.
func (e StateResponseStateInstance) Valid() bool {
	switch e {
	case Authorized:
		return true
	case Blocked:
		return true
	case NotAuthorized:
		return true
	case PendingPassword:
		return true
	case Starting:
		return true
	case Suspended:
		return true
	default:
		return false
	}
}

// Defines values for YesNo.
const (
	No  YesNo = "no"
	Yes YesNo = "yes"
)

// Valid indicates whether the value is a known member of the YesNo enum.
func (e YesNo) Valid() bool {
	switch e {
	case No:
		return true
	case Yes:
		return true
	default:
		return false
	}
}

// AccountSettingsResponse defines model for AccountSettingsResponse.
type AccountSettingsResponse struct {
	Avatar *string `json:"avatar,omitempty"`
	ChatId *string `json:"chatId,omitempty"`
	Phone  *string `json:"phone,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Code    string `json:"code"`
	Details string `json:"details"`
}

// SendFileResponse defines model for SendFileResponse.
type SendFileResponse struct {
	IdMessage *string `json:"idMessage,omitempty"`
}

// SendMessageResponse defines model for SendMessageResponse.
type SendMessageResponse struct {
	IdMessage *string `json:"idMessage,omitempty"`
}

// SettingsResponse defines model for SettingsResponse.
type SettingsResponse struct {
	DelaySendMessagesMilliseconds     *int64  `json:"delaySendMessagesMilliseconds,omitempty"`
	IncomingWebhook                   *YesNo  `json:"incomingWebhook,omitempty"`
	MarkIncomingMessagesReaded        *YesNo  `json:"markIncomingMessagesReaded,omitempty"`
	MarkIncomingMessagesReadedOnReply *YesNo  `json:"markIncomingMessagesReadedOnReply,omitempty"`
	OutgoingAPIMessageWebhook         *YesNo  `json:"outgoingAPIMessageWebhook,omitempty"`
	OutgoingMessageWebhook            *YesNo  `json:"outgoingMessageWebhook,omitempty"`
	OutgoingWebhook                   *YesNo  `json:"outgoingWebhook,omitempty"`
	StateWebhook                      *YesNo  `json:"stateWebhook,omitempty"`
	TypeInstance                      *string `json:"typeInstance,omitempty"`
	WebhookUrl                        *string `json:"webhookUrl,omitempty"`
	WebhookUrlToken                   *string `json:"webhookUrlToken,omitempty"`
	Wid                               *string `json:"wid,omitempty"`
}

// StateResponse defines model for StateResponse.
type StateResponse struct {
	StateInstance *StateResponseStateInstance `json:"stateInstance,omitempty"`
}

// StateResponseStateInstance defines model for StateResponse.StateInstance.
type StateResponseStateInstance string

// YesNo defines model for YesNo.
type YesNo string

// ApiToken defines model for ApiToken.
type ApiToken = string

// InstanceId defines model for InstanceId.
type InstanceId = string

// GetAccountSettingsParams defines parameters for GetAccountSettings.
type GetAccountSettingsParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// GetInstanceSettingsParams defines parameters for GetInstanceSettings.
type GetInstanceSettingsParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// GetInstanceStateParams defines parameters for GetInstanceState.
type GetInstanceStateParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// SendFileJSONBody defines parameters for SendFile.
type SendFileJSONBody struct {
	Caption  *string `json:"caption,omitempty"`
	ChatId   string  `json:"chatId"`
	FileName string  `json:"fileName"`
	UrlFile  string  `json:"urlFile"`
}

// SendFileParams defines parameters for SendFile.
type SendFileParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// SendMessageJSONBody defines parameters for SendMessage.
type SendMessageJSONBody struct {
	ChatId  string `json:"chatId"`
	Message string `json:"message"`
}

// SendMessageParams defines parameters for SendMessage.
type SendMessageParams struct {
	InstanceId InstanceId `form:"instance_id" json:"instance_id"`
	ApiToken   ApiToken   `form:"api_token" json:"api_token"`
}

// SendFileJSONRequestBody defines body for SendFile for application/json ContentType.
type SendFileJSONRequestBody SendFileJSONBody

// SendMessageJSONRequestBody defines body for SendMessage for application/json ContentType.
type SendMessageJSONRequestBody SendMessageJSONBody